Remove unused maps, create a new one, modify others
Remove specially the ones I don't remember and the ones I started to
use for execute expressions or navigate among tag find results, but
I stopped using it for a long time.

Create a new one that is useful for revision: to find an entire line
inside the huge journal file.

Modify the encoding one because <C-E><C-F> is not easy to remember. The
new one <C-E><C-E> looks more promising.